Исправлен баг, при обновлении состава группы, существующий студент не удалялся из таблицы student_group_stories

+++ app/Http/Controllers/Admin/Groups/GroupController.php
@@ -260,7 +260,6 @@ class GroupController extends Controller
                                 'name' => $studentPartFullName[1],
                                 'group_id' => $group->id,
                                 'user_id' => $user->id,
-                                'status' => 1
                             ]);
 
                             $this->studentGroupStory->create([
@@ -268,18 +267,20 @@ class GroupController extends Controller
                                 'group_story_id' => $groupStory->id
                             ]);
                         } else {
-                            $existStudent->update([
-                                'group_id' => $group->id,
-                            ]);
+                            if ($existStudent->group()->first()->id != $group->id) {
+                                $existStudent->group()->first()->groupStories()
+                                    ->where('year_history', '=', $group->year)->first()->studentGroupStory()
+                                    ->where('student_id', '=', $existStudent->id)->delete();
 
-                            $existStudent->group()->first()->groupStories()
-                                ->where('year_history', '=', $group->year)->first()->studentGroupStory()
-                                ->where('student_id', '=', $existStudent->id)->delete();
+                                $existStudent->update([
+                                    'group_id' => $group->id,
+                                ]);
 
-                            $this->studentGroupStory->create([
-                                'student_id' => $existStudent->id,
-                                'group_story_id' => $groupStory->id
-                            ]);
+                                $this->studentGroupStory->create([
+                                    'student_id' => $existStudent->id,
+                                    'group_story_id' => $groupStory->id
+                                ]);
+                            }
                         }
                     }
                     fclose($handle);
